Version 0.2
Added significant changes made by Benjamin Fritz. Now deletes the temporary folders it creates. the clearLint() function now only deletes the MLint matches rather than all of them as it had previously. Known bug: on unloading the buffer the temporary directory is deleted, reloading the buffer causes the script to error as it can't find the directory. I am working on a fix for this and should upload it soon. Also note that the script will on work with Vim 7.1 or higher.